Implement bytes-based LRU cache using Sizeable interface (#6722)
* Implement Sizeable interface usage to base LRU cache

+++ service/history/resource/resource.go
@@ -129,7 +129,7 @@ func New(
 		serviceResource.GetHistoryManager(),
 		params.Logger,
 		params.MetricsClient,
-		uint64(config.EventsCacheMaxSize()),
+		config.EventsCacheMaxSize,
 		serviceResource.GetDomainCache(),
 	)
 	ratelimitAlgorithm, err := algorithm.New(
